datavizai/backend/requirements.txt
flask
flask-cors
mysql-connector-python
scikit-learn
numpy
